move tests to test_core_model_loading.py
